- Moved failure mechanism extensions to data project. - Added missing tests for failure mechanism extensions. - Implemented getting the calculation beta in the failure mechanism extensions. Related to Issue [WTI-812]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@5458 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 8793824d56536e6857757f734dc7b605ce23d10b
Former-commit-id: de75f8055d4802bbfe9199214199f7f9608665a0

+++ Ringtoets/Revetment/src/Ringtoets.Revetment.Service/IWaveConditionsCalculationService.cs
@@ -60,7 +60,7 @@ namespace Ringtoets.Revetment.Service
                                        double a,
                                        double b,
                                        double c,
-                                       int norm,
+                                       double norm,
                                        WaveConditionsInput input,
                                        string hlcdDirectory,
                                        string ringId,
